changed Long variables to objects to all of the models

+++ src/main/java/com/revature/kyubbs/repositories/ReportRepository.java
@@ -8,8 +8,8 @@ public interface ReportRepository extends JpaRepository<Report, Long> {
 
 	public Report findReportByIp(String ip);
 
-	public Report findReportByPostId(Long id);
+	public Report findReportByPostId(int id);
 
-	public Report findReportByThreadId(Long id);
+	public Report findReportByThreadId(int id);
 
 }
